lib: sbi: memregion: Make memregions keep track of raw size rather than order
In order to accomodate isolation primitives other than PMP (such as the
forthcoming SMMTT implementation), make memregions keep track of their raw size
rather than prerounding the size to a power-of-two order. Not all isolation
primitives have the same stringent naturally-aligned power-of-two requirements
as PMP, and so we need a layer of abstraction between memregion information as
it is discovered (from device trees or otherwise) and isolation information as
applied using a specific primitive. This commit lays the groundwork for such an
abstraction.